#80 feat: s3 — league surface split (9 child routes)
Gut LeagueComponent's activeTab switch into nine standalone lazy-loaded
child components mounted under /league/... (Standings, Matchups, Playoffs,
World Cup, Rulebook, Scoring, League Settings, Payouts, Rule Proposals).

- Scaffold 9 standalone components, each owning its own template + SCSS
- LeagueComponent reduced to shell: header + router-outlet + data bootstrap
- Rewire sidebar.entries.ts to /league/<sub> flat routes (no queryParams)
- Add nested routes in app-routing.module.ts with loadComponent; /league
  redirects to /league/standings
- /selected-league stays single-view (Q1 hybrid — unchanged)
- npm run build clean; all 9 surfaces appear as lazy chunks
- Initial bundle shrinks from 930kB to 900kB vs baseline
