+++ src/app/pages/league/rules/league-settings/league-settings.component.html
@@ -1,0 +1,45 @@
+<!-- Roster Configuration -->
+<div class="rules-section">
+  <h2 class="rules-section-title">Roster Slots</h2>
+  <div class="roster-slots">
+    <div
+      class="roster-slot"
+      *ngFor="let slot of rosterSlots"
+      [attr.data-pos]="slot.position"
+    >
+      <span class="slot-position">{{ slot.position === 'SUPER_FLEX' ? 'SF' : slot.position }}</span>
+      <span class="slot-count" *ngIf="slot.count > 1">x{{ slot.count }}</span>
+    </div>
+  </div>
+</div>
+
+<!-- League Settings -->
+<div class="rules-section" *ngIf="league">
+  <h2 class="rules-section-title">League Settings</h2>
+  <div class="league-settings-grid">
+    <div class="setting-card">
+      <span class="setting-label">Format</span>
+      <span class="setting-value">{{ league.isDynasty() ? 'Dynasty' : 'Redraft' }}</span>
+    </div>
+    <div class="setting-card">
+      <span class="setting-label">Teams</span>
+      <span class="setting-value">{{ league.total_rosters || 12 }}</span>
+    </div>
+    <div class="setting-card">
+      <span class="setting-label">Playoff Teams</span>
+      <span class="setting-value">{{ league.getPlayoffTeams() }}</span>
+    </div>
+    <div class="setting-card">
+      <span class="setting-label">Taxi Slots</span>
+      <span class="setting-value">{{ league.getTaxiSlots() }}</span>
+    </div>
+    <div class="setting-card">
+      <span class="setting-label">Trade Deadline</span>
+      <span class="setting-value">Week {{ league.settings?.trade_deadline || 'N/A' }}</span>
+    </div>
+    <div class="setting-card">
+      <span class="setting-label">Divisions</span>
+      <span class="setting-value">{{ league.getDivisions()?.length || 0 }}</span>
+    </div>
+  </div>
+</div>
