+++ src/app/pages/league/rules/league-settings/league-settings.component.scss
@@ -1,0 +1,93 @@
+@import 'styles/variables';
+@import 'styles/mixins';
+
+.rules-section {
+  margin-bottom: $spacing-xl;
+}
+
+.rules-section-title {
+  font-family: $font-display;
+  font-size: 1.5rem;
+  letter-spacing: 0.05em;
+  color: $champion-gold;
+  margin: 0 0 $spacing-md;
+}
+
+// Roster Slots
+.roster-slots {
+  display: flex;
+  flex-wrap: wrap;
+  gap: $spacing-sm;
+}
+
+.roster-slot {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  padding: $spacing-sm $spacing-md;
+  background: rgba($dark-navy, 0.5);
+  border: 1px solid rgba($surface-light, 0.3);
+  border-radius: $radius-md;
+
+  &[data-pos="QB"] { border-left: 3px solid #5ba3ff; }
+  &[data-pos="RB"] { border-left: 3px solid $success-green; }
+  &[data-pos="WR"] { border-left: 3px solid $champion-gold; }
+  &[data-pos="TE"] { border-left: 3px solid #ff8a65; }
+  &[data-pos="FLEX"] { border-left: 3px solid $steel-blue; }
+  &[data-pos="SUPER_FLEX"] { border-left: 3px solid #ce93d8; }
+  &[data-pos="K"] { border-left: 3px solid $text-secondary; }
+  &[data-pos="DEF"] { border-left: 3px solid $accent-red; }
+  &[data-pos="BN"] { border-left: 3px solid $text-muted; }
+}
+
+.slot-position {
+  font-family: $font-mono;
+  font-size: $text-sm;
+  font-weight: 700;
+  color: $text-primary;
+}
+
+.slot-count {
+  font-family: $font-mono;
+  font-size: $text-xs;
+  color: $text-secondary;
+}
+
+// League Settings Grid
+.league-settings-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(160px, 1fr));
+  gap: $spacing-md;
+}
+
+.setting-card {
+  display: flex;
+  flex-direction: column;
+  gap: $spacing-xs;
+  padding: $spacing-md;
+  background: rgba($dark-navy, 0.5);
+  border: 1px solid rgba($surface-light, 0.2);
+  border-radius: $radius-md;
+}
+
+.setting-label {
+  font-size: $text-xs;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+  color: $text-secondary;
+}
+
+.setting-value {
+  font-family: $font-mono;
+  font-size: $text-lg;
+  font-weight: 700;
+  color: $text-primary;
+}
+
+// Mobile
+@media (max-width: 768px) {
+  .league-settings-grid {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
