+++ src/app/pages/league/rules/payouts/payouts.component.html
@@ -1,0 +1,4 @@
+<div class="rules-section">
+  <h2 class="rules-section-title">Dues &amp; Payouts</h2>
+  <div class="payouts-content" [innerHTML]="content"></div>
+</div>
