+++ src/app/pages/league/rules/rule-proposals/rule-proposals.component.html
@@ -1,0 +1,130 @@
+<div class="rules-section">
+  <div class="proposals-header">
+    <h2 class="rules-section-title">Rule Proposals</h2>
+    <span class="threshold-info">{{ approvalThreshold }}/{{ totalRosters }} votes to approve</span>
+    <button class="propose-btn" (click)="showProposalForm = !showProposalForm">
+      {{ showProposalForm ? 'Cancel' : 'Propose a Rule' }}
+    </button>
+  </div>
+
+  <!-- Proposal Form -->
+  <div class="proposal-form" *ngIf="showProposalForm">
+    <input
+      type="text"
+      class="proposal-input"
+      placeholder="Rule title..."
+      [(ngModel)]="proposalTitle"
+      maxlength="100"
+    />
+    <textarea
+      class="proposal-textarea"
+      placeholder="Describe the rule change..."
+      [(ngModel)]="proposalDescription"
+      rows="3"
+    ></textarea>
+    <button
+      class="submit-proposal-btn"
+      (click)="submitProposal()"
+      [disabled]="submittingProposal || !proposalTitle.trim()"
+    >
+      {{ submittingProposal ? 'Submitting...' : 'Submit Proposal' }}
+    </button>
+  </div>
+
+  <!-- Filter Buttons -->
+  <div class="proposal-filters" *ngIf="proposals.length > 0">
+    <button class="filter-btn" [class.active]="proposalFilter === 'all'" (click)="proposalFilter = 'all'">All</button>
+    <button class="filter-btn" [class.active]="proposalFilter === 'open'" (click)="proposalFilter = 'open'">Open</button>
+    <button class="filter-btn filter-approved" [class.active]="proposalFilter === 'approved'" (click)="proposalFilter = 'approved'">Approved</button>
+    <button class="filter-btn filter-denied" [class.active]="proposalFilter === 'rejected'" (click)="proposalFilter = 'rejected'">Denied</button>
+  </div>
+
+  <!-- Proposals List -->
+  <div class="proposals-list" *ngIf="filteredProposals.length > 0">
+    <div
+      class="proposal-card"
+      *ngFor="let proposal of filteredProposals"
+      [class.approved]="proposal.status === 'approved'"
+      [class.rejected]="proposal.status === 'rejected'"
+      [class.closed]="proposal.status === 'closed'"
+    >
+      <div
+        class="stamp-overlay"
+        *ngIf="proposal.status === 'approved' || proposal.status === 'rejected'"
+        [class.stamp-approved]="proposal.status === 'approved'"
+        [class.stamp-denied]="proposal.status === 'rejected'"
+        [class.stamp-animate]="recentlyStamped.has(proposal.id)"
+      >
+        {{ proposal.status === 'approved' ? 'APPROVED' : 'DENIED' }}
+      </div>
+
+      <div class="proposal-top">
+        <span
+          class="proposal-status"
+          [class.status-open]="proposal.status === 'open'"
+          [class.status-approved]="proposal.status === 'approved'"
+          [class.status-rejected]="proposal.status === 'rejected'"
+          [class.status-closed]="proposal.status === 'closed'"
+        >
+          {{ proposal.status }}
+        </span>
+        <div class="proposal-top-right">
+          <span class="proposal-date">{{ getProposalDate(proposal.created_at) }}</span>
+          <button
+            class="delete-proposal-btn"
+            *ngIf="proposal.proposed_by === currentUserId && proposal.status === 'open'"
+            (click)="deleteProposal(proposal.id)"
+            title="Delete proposal"
+          >
+            &times;
+          </button>
+        </div>
+      </div>
+      <h4 class="proposal-title">{{ proposal.title }}</h4>
+      <p class="proposal-description" *ngIf="proposal.description">{{ proposal.description }}</p>
+      <span class="proposal-author">Proposed by {{ proposal.proposed_by_username }}</span>
+
+      <div class="vote-progress" *ngIf="proposal.yes_count + proposal.no_count > 0">
+        <div class="vote-bar">
+          <div
+            class="vote-bar-yes"
+            [style.width.%]="(proposal.yes_count / totalRosters) * 100"
+          ></div>
+          <div
+            class="vote-bar-no"
+            [style.width.%]="(proposal.no_count / totalRosters) * 100"
+          ></div>
+        </div>
+        <div class="vote-bar-labels">
+          <span class="vote-bar-label yes">{{ proposal.yes_count }} Yes</span>
+          <span class="vote-bar-label no">{{ proposal.no_count }} No</span>
+        </div>
+      </div>
+
+      <div class="vote-section" *ngIf="proposal.status === 'open'">
+        <button
+          class="vote-btn yes"
+          [class.voted]="proposal.my_vote === 'yes'"
+          (click)="castVote(proposal.id, 'yes')"
+        >
+          {{ proposal.my_vote === 'yes' ? 'Voted Yes' : 'Vote Yes' }}
+        </button>
+        <button
+          class="vote-btn no"
+          [class.voted]="proposal.my_vote === 'no'"
+          (click)="castVote(proposal.id, 'no')"
+        >
+          {{ proposal.my_vote === 'no' ? 'Voted No' : 'Vote No' }}
+        </button>
+      </div>
+    </div>
+  </div>
+
+  <p class="no-proposals" *ngIf="proposals.length > 0 && filteredProposals.length === 0">
+    No {{ proposalFilter }} proposals.
+  </p>
+
+  <p class="no-proposals" *ngIf="proposals.length === 0">
+    No proposals yet. Be the first to suggest a rule change!
+  </p>
+</div>
