+++ src/app/pages/league/rules/rule-proposals/rule-proposals.component.scss
@@ -1,0 +1,409 @@
+@import 'styles/variables';
+@import 'styles/mixins';
+
+.rules-section {
+  margin-bottom: $spacing-xl;
+}
+
+.rules-section-title {
+  font-family: $font-display;
+  font-size: 1.5rem;
+  letter-spacing: 0.05em;
+  color: $champion-gold;
+  margin: 0 0 $spacing-md;
+}
+
+// Proposals
+.proposals-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: $spacing-md;
+
+  .rules-section-title {
+    margin: 0;
+  }
+}
+
+.propose-btn {
+  background: transparent;
+  color: $champion-gold;
+  border: 1px solid $champion-gold;
+  border-radius: $radius-md;
+  padding: 8px $spacing-md;
+  font-size: $text-sm;
+  font-weight: 600;
+  font-family: $font-body;
+  cursor: pointer;
+  transition: all $transition-base;
+
+  &:hover {
+    background: rgba($champion-gold, 0.1);
+  }
+}
+
+.proposal-form {
+  display: flex;
+  flex-direction: column;
+  gap: $spacing-sm;
+  padding: $spacing-md;
+  background: rgba($dark-navy, 0.5);
+  border: 1px solid rgba($surface-light, 0.3);
+  border-radius: $radius-lg;
+  margin-bottom: $spacing-lg;
+}
+
+.proposal-input,
+.proposal-textarea {
+  background: $bg-input;
+  color: $text-primary;
+  border: 1px solid rgba($surface-light, 0.3);
+  border-radius: $radius-md;
+  padding: 10px $spacing-md;
+  font-size: $text-sm;
+  font-family: $font-body;
+  outline: none;
+  transition: border-color $transition-base;
+
+  &:focus {
+    border-color: $champion-gold;
+  }
+
+  &::placeholder {
+    color: $text-muted;
+  }
+}
+
+.proposal-textarea {
+  resize: vertical;
+  min-height: 60px;
+}
+
+.submit-proposal-btn {
+  align-self: flex-end;
+  background: $gold-accent;
+  color: $deep-navy;
+  border: none;
+  border-radius: $radius-md;
+  padding: 10px $spacing-lg;
+  font-size: $text-sm;
+  font-weight: 700;
+  font-family: $font-body;
+  cursor: pointer;
+  transition: all $transition-base;
+
+  &:hover:not(:disabled) {
+    filter: brightness(1.1);
+  }
+
+  &:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+  }
+}
+
+.proposal-filters {
+  display: flex;
+  gap: 6px;
+  margin-bottom: $spacing-md;
+}
+
+.filter-btn {
+  background: rgba($dark-navy, 0.6);
+  color: $text-secondary;
+  border: 1px solid rgba($surface-light, 0.3);
+  border-radius: $radius-xl;
+  padding: 6px 16px;
+  font-size: $text-xs;
+  font-weight: 600;
+  font-family: $font-body;
+  cursor: pointer;
+  transition: all $transition-base;
+
+  &:hover:not(.active) {
+    background: rgba($surface-light, 0.2);
+    color: $text-primary;
+  }
+
+  &.active {
+    background: $gold-accent;
+    color: $deep-navy;
+    border-color: $champion-gold;
+  }
+
+  &.filter-approved.active {
+    background: rgba($success-green, 0.2);
+    color: $success-green;
+    border-color: $success-green;
+  }
+
+  &.filter-denied.active {
+    background: rgba($error-red, 0.2);
+    color: $error-red;
+    border-color: $error-red;
+  }
+}
+
+.proposals-list {
+  display: flex;
+  flex-direction: column;
+  gap: $spacing-md;
+}
+
+.proposal-card {
+  position: relative;
+  overflow: hidden;
+  padding: $spacing-md;
+  background: rgba($dark-navy, 0.5);
+  border: 1px solid rgba($surface-light, 0.2);
+  border-radius: $radius-lg;
+  transition: all $transition-base;
+
+  &.approved {
+    border-left: 3px solid $success-green;
+  }
+
+  &.rejected {
+    border-left: 3px solid $error-red;
+    opacity: 0.85;
+  }
+
+  &.closed {
+    opacity: 0.5;
+  }
+}
+
+// Stamp overlay
+.stamp-overlay {
+  position: absolute;
+  top: 50%;
+  right: 16px;
+  transform: translateY(-50%) rotate(-12deg);
+  font-size: 2rem;
+  font-weight: 900;
+  font-family: $font-display;
+  letter-spacing: 0.1em;
+  text-transform: uppercase;
+  padding: 6px 18px;
+  border-radius: $radius-sm;
+  border: 4px solid;
+  pointer-events: none;
+  opacity: 0.25;
+  z-index: 1;
+
+  &.stamp-approved {
+    color: $success-green;
+    border-color: $success-green;
+  }
+
+  &.stamp-denied {
+    color: $error-red;
+    border-color: $error-red;
+  }
+
+  &.stamp-animate {
+    animation: stampSlam 0.5s cubic-bezier(0.2, 0.8, 0.3, 1) forwards;
+  }
+}
+
+@keyframes stampSlam {
+  0% {
+    transform: translateY(-50%) rotate(-12deg) scale(2.5);
+    opacity: 0;
+  }
+  50% {
+    opacity: 0.4;
+  }
+  80% {
+    transform: translateY(-50%) rotate(-12deg) scale(0.95);
+  }
+  100% {
+    transform: translateY(-50%) rotate(-12deg) scale(1);
+    opacity: 0.25;
+  }
+}
+
+.threshold-info {
+  font-size: $text-xs;
+  color: $text-muted;
+  font-family: $font-mono;
+}
+
+.proposal-top {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: $spacing-xs;
+}
+
+.proposal-status {
+  font-size: $text-xs;
+  font-weight: 700;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+  padding: 2px 8px;
+  border-radius: $radius-sm;
+
+  &.status-open {
+    color: $champion-gold;
+    background: rgba($champion-gold, 0.12);
+  }
+
+  &.status-approved {
+    color: $success-green;
+    background: rgba($success-green, 0.12);
+  }
+
+  &.status-rejected {
+    color: $error-red;
+    background: rgba($error-red, 0.12);
+  }
+
+  &.status-closed {
+    color: $text-muted;
+    background: rgba($text-muted, 0.12);
+  }
+}
+
+.proposal-top-right {
+  display: flex;
+  align-items: center;
+  gap: $spacing-xs;
+}
+
+.proposal-date {
+  font-size: $text-xs;
+  color: $text-muted;
+}
+
+.delete-proposal-btn {
+  background: none;
+  border: none;
+  color: $text-muted;
+  font-size: 1.1rem;
+  cursor: pointer;
+  padding: 0 4px;
+  line-height: 1;
+  opacity: 0.6;
+  transition: opacity 0.2s, color 0.2s;
+
+  &:hover {
+    opacity: 1;
+    color: $error-red;
+  }
+}
+
+.proposal-title {
+  font-size: $text-base;
+  font-weight: 700;
+  color: $text-primary;
+  margin: 0 0 $spacing-xs;
+}
+
+.proposal-description {
+  font-size: $text-sm;
+  color: $text-secondary;
+  margin: 0 0 $spacing-sm;
+  line-height: 1.5;
+}
+
+.proposal-author {
+  font-size: $text-xs;
+  color: $text-muted;
+}
+
+.vote-section {
+  display: flex;
+  gap: $spacing-sm;
+  margin-top: $spacing-sm;
+}
+
+.vote-btn {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  padding: 6px $spacing-md;
+  border-radius: $radius-md;
+  font-size: $text-sm;
+  font-weight: 600;
+  font-family: $font-body;
+  cursor: pointer;
+  transition: all $transition-base;
+  background: transparent;
+
+  &.yes {
+    color: $success-green;
+    border: 1px solid rgba($success-green, 0.3);
+
+    &:hover,
+    &.voted {
+      background: rgba($success-green, 0.15);
+      border-color: $success-green;
+    }
+  }
+
+  &.no {
+    color: $error-red;
+    border: 1px solid rgba($error-red, 0.3);
+
+    &:hover,
+    &.voted {
+      background: rgba($error-red, 0.15);
+      border-color: $error-red;
+    }
+  }
+}
+
+// Vote progress bar
+.vote-progress {
+  margin-top: $spacing-sm;
+}
+
+.vote-bar {
+  display: flex;
+  height: 6px;
+  border-radius: 3px;
+  background: rgba($surface-light, 0.15);
+  overflow: hidden;
+}
+
+.vote-bar-yes {
+  background: $success-green;
+  transition: width 0.4s ease;
+}
+
+.vote-bar-no {
+  background: $error-red;
+  margin-left: auto;
+  transition: width 0.4s ease;
+}
+
+.vote-bar-labels {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 4px;
+}
+
+.vote-bar-label {
+  font-size: $text-xs;
+  font-family: $font-mono;
+
+  &.yes { color: $success-green; }
+  &.no { color: $error-red; }
+}
+
+.no-proposals {
+  text-align: center;
+  color: $text-muted;
+  font-size: $text-sm;
+  padding: $spacing-xl 0;
+}
+
+// Mobile
+@media (max-width: 768px) {
+  .proposals-header {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: $spacing-sm;
+  }
+}
