+++ src/app/pages/league/rules/rulebook/rulebook.component.html
@@ -1,0 +1,16 @@
+<div class="rules-section">
+  <h2 class="rules-section-title">League Rulebook</h2>
+  <div class="rulebook">
+    <div
+      class="rulebook-chapter"
+      *ngFor="let rule of LEAGUE_RULES; let i = index"
+      [class.expanded]="expandedRuleSections.has(i)"
+    >
+      <button class="chapter-header" (click)="toggleRuleSection(i)">
+        <span class="chapter-title">{{ rule.title }}</span>
+        <span class="chapter-arrow">{{ expandedRuleSections.has(i) ? '&#9650;' : '&#9660;' }}</span>
+      </button>
+      <div class="chapter-content" *ngIf="expandedRuleSections.has(i)" [innerHTML]="rule.content"></div>
+    </div>
+  </div>
+</div>
