+++ src/app/pages/league/rules/rulebook/rulebook.component.scss
@@ -1,0 +1,109 @@
+@import 'styles/variables';
+@import 'styles/mixins';
+
+.rules-section {
+  margin-bottom: $spacing-xl;
+}
+
+.rules-section-title {
+  font-family: $font-display;
+  font-size: 1.5rem;
+  letter-spacing: 0.05em;
+  color: $champion-gold;
+  margin: 0 0 $spacing-md;
+}
+
+// Rulebook
+.rulebook {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+}
+
+.rulebook-chapter {
+  background: rgba($dark-navy, 0.4);
+  border: 1px solid rgba($surface-light, 0.12);
+  border-radius: $radius-md;
+  overflow: hidden;
+  transition: all $transition-base;
+
+  &.expanded {
+    border-color: rgba($champion-gold, 0.3);
+  }
+}
+
+.chapter-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  padding: $spacing-sm $spacing-md;
+  background: none;
+  border: none;
+  color: $text-primary;
+  font-size: $text-sm;
+  font-weight: 600;
+  font-family: $font-body;
+  cursor: pointer;
+  transition: background $transition-base;
+
+  &:hover {
+    background: rgba($surface-light, 0.06);
+  }
+}
+
+.chapter-title {
+  text-align: left;
+}
+
+.chapter-arrow {
+  font-size: $text-xs;
+  color: $text-muted;
+}
+
+.chapter-content {
+  padding: 0 $spacing-md $spacing-md;
+  font-size: $text-sm;
+  color: $text-secondary;
+  line-height: 1.7;
+  animation: fadeIn 0.2s ease;
+
+  strong {
+    color: $text-primary;
+  }
+
+  em {
+    color: $champion-gold;
+  }
+}
+
+@keyframes fadeIn {
+  from { opacity: 0; transform: translateY(-4px); }
+  to { opacity: 1; transform: translateY(0); }
+}
+
+:host ::ng-deep .rules-table {
+  width: 100%;
+  margin: $spacing-xs 0;
+  border-collapse: collapse;
+  font-size: $text-xs;
+
+  th {
+    text-align: left;
+    padding: 6px 10px;
+    background: rgba($surface-light, 0.08);
+    color: $champion-gold;
+    font-weight: 600;
+    border-bottom: 1px solid rgba($surface-light, 0.15);
+  }
+
+  td {
+    padding: 5px 10px;
+    border-bottom: 1px solid rgba($surface-light, 0.06);
+    color: $text-secondary;
+  }
+
+  tr:last-child td {
+    border-bottom: none;
+  }
+}
