+++ src/app/pages/league/rules/scoring/scoring.component.scss
@@ -1,0 +1,77 @@
+@import 'styles/variables';
+@import 'styles/mixins';
+
+.rules-section {
+  margin-bottom: $spacing-xl;
+}
+
+.rules-section-title {
+  font-family: $font-display;
+  font-size: 1.5rem;
+  letter-spacing: 0.05em;
+  color: $champion-gold;
+  margin: 0 0 $spacing-md;
+}
+
+// Scoring Settings
+.scoring-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(260px, 1fr));
+  gap: $spacing-md;
+}
+
+.scoring-category {
+  background: rgba($dark-navy, 0.5);
+  border: 1px solid rgba($surface-light, 0.2);
+  border-radius: $radius-lg;
+  overflow: hidden;
+}
+
+.scoring-category-name {
+  font-family: $font-display;
+  font-size: 1.1rem;
+  letter-spacing: 0.05em;
+  color: $text-primary;
+  padding: $spacing-sm $spacing-md;
+  margin: 0;
+  background: rgba($champion-gold, 0.08);
+  border-bottom: 1px solid rgba($surface-light, 0.2);
+}
+
+.scoring-row {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 8px $spacing-md;
+  border-bottom: 1px solid rgba($surface-light, 0.1);
+
+  &:last-child {
+    border-bottom: none;
+  }
+}
+
+.scoring-label {
+  font-size: $text-sm;
+  color: $text-secondary;
+}
+
+.scoring-value {
+  font-family: $font-mono;
+  font-size: $text-sm;
+  font-weight: 700;
+
+  &.positive {
+    color: $champion-gold;
+  }
+
+  &.negative {
+    color: $error-red;
+  }
+}
+
+// Mobile
+@media (max-width: 768px) {
+  .scoring-grid {
+    grid-template-columns: 1fr;
+  }
+}
