+++ src/app/pages/league/world-cup/world-cup.component.html
@@ -1,0 +1,60 @@
+<app-loader [loading]="loading"></app-loader>
+
+<div class="worldcup-container" *ngIf="worldCupLoaded">
+  <div class="worldcup-header">
+    <h2 class="worldcup-title">World Cup Qualifying</h2>
+    <p class="worldcup-subtitle">
+      Divisional head-to-head records across {{ worldCupSeasons.length }} season{{ worldCupSeasons.length !== 1 ? 's' : '' }}
+      ({{ worldCupSeasons.join(', ') }})
+    </p>
+    <p class="worldcup-rule">Top 2 in each division qualify</p>
+  </div>
+
+  <div class="worldcup-divisions">
+    <div class="wc-division" *ngFor="let division of worldCupDivisions">
+      <div class="wc-division-header">
+        <h3 class="wc-division-name">{{ division.divisionName }}</h3>
+      </div>
+
+      <div class="wc-standings-header" [style.grid-template-columns]="wcGridColumns">
+        <span class="wc-rank">#</span>
+        <span class="wc-team">Team</span>
+        <span class="wc-record">W</span>
+        <span class="wc-record">L</span>
+        <span class="wc-pts">PF</span>
+        <span class="wc-pts">PA</span>
+        <span class="wc-seasons" *ngFor="let season of worldCupSeasons">{{ season }}</span>
+      </div>
+
+      <div
+        class="wc-standings-row"
+        *ngFor="let team of division.teams; let i = index"
+        [class.qualified]="team.qualified"
+        [class.eliminated]="!team.qualified"
+        [style.grid-template-columns]="wcGridColumns"
+      >
+        <span class="wc-rank" [class.gold]="i === 0" [class.silver]="i === 1">
+          {{ i + 1 }}
+        </span>
+        <div class="wc-team">
+          <span class="wc-team-name">{{ team.team_name || team.username }}</span>
+          <span class="wc-username">{{ team.username }}</span>
+        </div>
+        <span class="wc-record wins">{{ team.wins }}</span>
+        <span class="wc-record losses">{{ team.losses }}</span>
+        <span class="wc-pts">{{ team.pointsFor.toFixed(1) }}</span>
+        <span class="wc-pts">{{ team.pointsAgainst.toFixed(1) }}</span>
+        <span class="wc-seasons" *ngFor="let season of worldCupSeasons">
+          <ng-container *ngIf="getSeasonBreakdown(team, season) as sb">
+            {{ sb.wins }}-{{ sb.losses }}
+          </ng-container>
+        </span>
+      </div>
+
+      <div class="wc-qualify-line" *ngIf="division.teams.length > 2">
+        <span class="qualify-tag">Qualified</span>
+        <span class="qualify-divider"></span>
+      </div>
+    </div>
+  </div>
+</div>
