+++ src/app/pages/league/world-cup/world-cup.component.scss
@@ -1,0 +1,194 @@
+@import 'styles/variables';
+@import 'styles/mixins';
+
+.worldcup-container {
+  padding: $spacing-md 0;
+}
+
+.worldcup-header {
+  text-align: center;
+  margin-bottom: $spacing-xl;
+}
+
+.worldcup-title {
+  font-family: $font-display;
+  font-size: 2rem;
+  letter-spacing: 0.05em;
+  color: $champion-gold;
+  margin: 0;
+}
+
+.worldcup-subtitle {
+  font-size: $text-base;
+  color: $text-secondary;
+  margin: $spacing-xs 0 0;
+}
+
+.worldcup-rule {
+  font-size: $text-sm;
+  color: $success-green;
+  font-weight: 600;
+  margin-top: $spacing-xs;
+}
+
+.worldcup-divisions {
+  display: flex;
+  flex-direction: column;
+  gap: $spacing-xl;
+}
+
+.wc-division {
+  background: rgba($dark-navy, 0.4);
+  border: 1px solid rgba($surface-light, 0.2);
+  border-radius: $radius-lg;
+  padding: $spacing-lg;
+  position: relative;
+}
+
+.wc-division-header {
+  text-align: center;
+  margin-bottom: $spacing-md;
+}
+
+.wc-division-name {
+  font-family: $font-display;
+  font-size: 1.5rem;
+  letter-spacing: 0.05em;
+  color: $champion-gold;
+  margin: 0;
+}
+
+.wc-standings-header,
+.wc-standings-row {
+  display: grid;
+  grid-template-columns: 40px 2fr 0.6fr 0.6fr 1fr 1fr repeat(auto-fill, 0.8fr);
+  gap: $spacing-sm;
+  align-items: center;
+  padding: 10px $spacing-md;
+}
+
+.wc-standings-header {
+  font-weight: 600;
+  font-size: $text-xs;
+  color: $text-secondary;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+  border-bottom: 1px solid rgba($surface-light, 0.3);
+}
+
+.wc-standings-row {
+  border-radius: $radius-md;
+  margin-bottom: $spacing-xs;
+  transition: all $transition-base;
+
+  &.qualified {
+    background: rgba($success-green, 0.08);
+    border-left: 3px solid $success-green;
+  }
+
+  &.eliminated {
+    opacity: 0.6;
+  }
+}
+
+.wc-rank {
+  text-align: center;
+  font-weight: 700;
+  font-family: $font-mono;
+  font-size: $text-base;
+
+  &.gold {
+    color: $champion-gold;
+    text-shadow: 0 0 8px rgba($champion-gold, 0.4);
+  }
+
+  &.silver {
+    color: silver;
+  }
+}
+
+.wc-team {
+  display: flex;
+  flex-direction: column;
+  min-width: 0;
+}
+
+.wc-team-name {
+  font-size: $text-sm;
+  font-weight: 600;
+  color: $text-primary;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.wc-username {
+  font-size: $text-xs;
+  color: $text-secondary;
+}
+
+.wc-record {
+  font-family: $font-mono;
+  font-weight: 700;
+  font-size: $text-base;
+  text-align: center;
+
+  &.wins {
+    color: $success-green;
+  }
+
+  &.losses {
+    color: $error-red;
+  }
+}
+
+.wc-pts {
+  font-family: $font-mono;
+  font-size: $text-sm;
+  color: $text-secondary;
+  text-align: center;
+}
+
+.wc-seasons {
+  font-family: $font-mono;
+  font-size: $text-sm;
+  color: $text-secondary;
+  text-align: center;
+}
+
+.wc-qualify-line {
+  display: flex;
+  align-items: center;
+  gap: $spacing-sm;
+  margin-top: $spacing-sm;
+  padding: 0 $spacing-md;
+
+  .qualify-tag {
+    font-size: $text-xs;
+    font-weight: 700;
+    color: $success-green;
+    text-transform: uppercase;
+    letter-spacing: 0.05em;
+    white-space: nowrap;
+  }
+
+  .qualify-divider {
+    flex: 1;
+    height: 1px;
+    background: rgba($success-green, 0.3);
+  }
+}
+
+// Mobile
+@media (max-width: 768px) {
+  .wc-standings-header,
+  .wc-standings-row {
+    grid-template-columns: 30px 1.5fr 0.5fr 0.5fr;
+    font-size: $text-xs;
+
+    .wc-pts,
+    .wc-seasons {
+      display: none;
+    }
+  }
+}
